setup.sh: Fix NIC hang issue on tx configuration
PTP out of sync once the DPDK talker and listener
start to run. It should able to sync again after that
but due to the queue congested, the NIC force to reset.
As a result, the system hang and require hard reboot.

This patch fix the taprio on tx side which causing the
queue congested with the packets.

Signed-off-by: Mohamad Noor Alim Hussin <mohamad.noor.alim.hussin@intel.com>