run.sh: Add "board" parameter
The "board" is refer to tgl or icx.
This parameter used to identify the configuration of the NIC
correspond to the board. As the same NIC might have difference
behavior on difference board.

To execute run.sh script on tgl with i255 NIC, the run command as follows
$ ./run i225 tgl <setup/run> <talker/listener> [APP component]

See README_run.md for the example of usage in detail.

Signed-off-by: Mohamad Noor Alim Hussin <mohamad.noor.alim.hussin@intel.com>